Use an int for the maximum object size to be cached in bytes since it is cached using a byte array and that is limited to Integer.MAX_VALUE. Test objects to be cached against the maximum object size, not the maximum cache size.
git-svn-id: https://svn.apache.org/repos/asf/tomcat/trunk@1549706 13f79535-47bb-0310-9956-ffa450edef68

+++ java/org/apache/catalina/WebResourceRoot.java
@@ -346,19 +346,21 @@ public interface WebResourceRoot extends Lifecycle {
     long getCacheMaxSize();
 
     /**
-     * Set the maximum permitted size for a single object in the cache.
+     * Set the maximum permitted size for a single object in the cache. Note
+     * that the maximum size in bytes may not exceed {@link Integer#MAX_VALUE}.
      *
      * @param cacheMaxObjectSize    Maximum size for a single cached object in
      *                              kilobytes
      */
-    void setCacheMaxObjectSize(long cacheMaxObjectSize);
+    void setCacheMaxObjectSize(int cacheMaxObjectSize);
 
     /**
-     * Get the maximum permitted size for a single object in the cache.
+     * Get the maximum permitted size for a single object in the cache. Note
+     * that the maximum size in bytes may not exceed {@link Integer#MAX_VALUE}.
      *
      * @return  Maximum size for a single cached object in kilobytes
      */
-    long getCacheMaxObjectSize();
+    int getCacheMaxObjectSize();
 
     /**
      * Controls whether the trace locked files feature is enabled. If enabled,
